Insert diddled, select Data methods in
Insert is now a method on zygote.Data - it also checks to see if the
client has been added before.

Also added a bunch of functions to pluck a zygote from the database.
Those're by Name, by ID and by Email. All which are set up as indexes in
the table. They all only return one zygote because that's only how many
should be in the database with that ID.

=)