Added madlib
madlib is a really simple templating format that uses HTML comments and
maps to render elements.

Some initial work on the client library is in, but it's still a work in
progress.

register page is still being figured out. Will have to write a module
separate from jam for emailing.